+++ wsgidav/lock_storage.py
@@ -10,6 +10,8 @@ Two alternative lock storage classes are defined here: one in-memory
 (dict-based), and one persistent low performance variant using shelve.
 
 See :class:`~wsgidav.lock_manager.LockManager`
+See :class:`~wsgidav.lock_storage.LockStorageDict`
+See :class:`~wsgidav.lock_storage.LockStorageShelve`
 """
 import os
 import shelve
@@ -54,7 +56,7 @@ class LockStorageDict:
     members is done by re-assignment and we call a _flush() method.
 
     This is obviously not persistent, but should be enough in some cases.
-    For a persistent implementation, see lock_manager.LockStorageShelve().
+    For a persistent implementation, see lock_storage.LockStorageShelve().
 
     Notes:
         expire is stored as expiration date in seconds since epoch (not in